WareHouse/SuperReader_liuyuan_20231204144309/requirements.txt
PyMuPDF==1.18.14  # Replace with the latest version number or the one you have tested with
ebooklib==0.17.1  # Replace with the latest version number or the one you have tested with
Pillow==8.3.2     # Replace with the latest version number or the one you have tested with
